+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+*
+
+### Changed
+
+*
+
+### Fixed
+
+*
+
+## [2.12.0] - 2021-12-05
+
+### Added
+
 * Add `localeToNative`, `localeToNativeP`, `localeToNativeB` and `localeToNativeBP` methods to convert locale sizes to native sizes and unit tests - [#4638](https://github.com/ripe-tech/ripe-core/issues/4638)
 * Add `Invoice Rules` API methods - [peri-invoicing/#3](https://github.com/ripe-tech/peri-invoicing/issues/3)
 * Add `textLengthP` and `textLength` to get the length of the value of the initials
@@ -17,10 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Made the `initialsBuilder` function `async` to support async calls in external initials builder logic
 * Added `ctx` to `initialsBuilder` call so that there is parity between the Javascript and Python initials builders
-
-### Fixed
-
-*
 
 ## [2.11.0] - 2021-12-03